+++ src/__test__/fixtures/types/User.ts
@@ -1,4 +1,4 @@
-import { PickFromPath } from '@konecty/sdk/TypeUtils';
+import { PickFromPath, UnionToIntersection } from '@konecty/sdk/TypeUtils';
 import {
 	KonectyModule,
 	ModuleConfig,
@@ -27,11 +27,11 @@ const userConfig: ModuleConfig = {
 };
 export type UserGroupType = { _id: string; name: string };
 export type UserGroupsType = { _id: string; name: string };
-export type UserRoleType = PickFromPath<Role, '_id' | 'name'>;
+export type UserRoleType = UnionToIntersection<PickFromPath<Role, '_id' | 'name'>>;
 export type UserCreatedByType = { _id: string; name: string; group: { name: unknown } };
 export type UserUpdatedByType = { _id: string; name: string; group: { name: unknown } };
 export type UserUserType = { _id: string; name: string; group: { name: unknown }; active: boolean };
-export type UserTargetQueueType = PickFromPath<Queue, '_id' | 'name'>;
+export type UserTargetQueueType = UnionToIntersection<PickFromPath<Queue, '_id' | 'name'>>;
 export type UserDirectorType = { _id: string; nickname: string };
 export type UserLocaleType = 'pt_BR' | 'en';
 export type UserStatusType = 'online' | 'away' | 'busy' | 'offline';